Uploaded finalized v1.2
finalized ver. 1.2 | progress to start the fully automation is key, and semi is... not so much. Next function will be a mega scrape for content under your active lists on your chosen account! Also, ive been running the script myself for an entire day, not a single ghost ban! progress!